+++ src/backend.rs
@@ -1,0 +1,45 @@
+use dioxus::prelude::*;
+
+
+// Query the database and return the last 10 dogs and their url
+#[server]
+pub async fn list_dogs() -> Result<Vec<(usize, String)>, ServerFnError> {
+    let dogs = DB.with(|f| {
+        f.prepare("SELECT id, url FROM dogs ORDER BY id DESC LIMIT 10")
+            .unwrap()
+            .query_map([], |row| Ok((row.get(0)?, row.get(1)?)))
+            .unwrap()
+            .map(|r| r.unwrap())
+            .collect()
+    });
+
+    Ok(dogs)
+}
+
+
+// Expose a `save_dog` endpoint on our server that takes an "image" parameter
+#[server]
+pub async fn save_dog(image: String) -> Result<(), ServerFnError> {
+    DB.with(|f| f.execute("INSERT INTO dogs (url) VALUES (?1)", &[&image]))?;
+    Ok(())
+}
+
+// The database is only xavailable to server code
+#[cfg(feature = "server")]
+thread_local! {
+    pub static DB: rusqlite::Connection = {
+        // Open the database from the persisted "hotdog.db" file
+        let conn = rusqlite::Connection::open("hotdog.db").expect("Failed to open database");
+
+        // Create the "dogs" table if it doesn't already exist
+        conn.execute_batch(
+            "CREATE TABLE IF NOT EXISTS dogs (
+                id INTEGER PRIMARY KEY,
+                url TEXT NOT NULL
+            );",
+        ).unwrap();
+
+        // Return the connection
+        conn
+    };
+}
