+++ src/components/nav.rs
@@ -1,0 +1,14 @@
+use crate::Route;
+use dioxus::prelude::*;
+#[component]
+pub fn NavBar() -> Element {
+    rsx! {
+        div { id: "title",
+            Link { to: Route::DogView,
+                h1 { "🌭 HotDogh! " }
+            }
+            Link { to: Route::Favorites, id: "heart", "♥️" } // <------- add this Link
+        }
+        Outlet::<Route> {}
+    }
+}
