+++ src/main.rs
@@ -1,7 +1,7 @@
 mod backend;
 mod components;
 
-use crate::components::*;
+use components::{DogView, NavBar, Favorites};
 use dioxus::prelude::*;
 
 fn main() {
@@ -23,93 +23,6 @@ fn app() -> Element {
         Router::<Route> {}
     }
 }
-#[component]
-fn Title() -> Element {
-    // Consume that type as a Context
-    let title = use_context::<TitleState>();
-    rsx! {
-        h1 { "{title.0}" }
-    }
-}
-
-
-#[component]
-pub fn NavBar() -> Element {
-    rsx! {
-        div { id: "title",
-            Link { to: Route::DogView,
-                h1 { "🌭 HotDog! " }
-            }
-            Link { to: Route::Favorites, id: "heart", "♥️" } // <------- add this Link
-        }
-        Outlet::<Route> {}
-    }
-}
-
-
-#[component]
-fn DogView() -> Element {
-    let mut img_src = use_resource(|| async move {
-        reqwest::get("https://dog.ceo/api/breeds/image/random")
-            .await
-            .unwrap()
-            .json::<DogApi>()
-            .await
-            .unwrap()
-            .message
-    });
-
-    rsx! {
-        div { id: "dogview",
-            img { src: img_src.cloned().unwrap_or_default() }
-        }
-        div { id: "buttons",
-            button { onclick: move |_| img_src.restart(), id: "skip", "skip" }
-                button {
-                id: "save",
-                onclick: move |_| async move {
-                    let current = img_src.cloned().unwrap();
-                    img_src.restart();
-                    _ = save_dog(current).await;
-                },
-
-                "save!"
-            }
-        }
-    }
-}
-#[derive(serde::Deserialize)]
-struct DogApi {
-    message: String,
-}
-
-// Expose a `save_dog` endpoint on our server that takes an "image" parameter
-#[server]
-pub async fn save_dog(image: String) -> Result<(), ServerFnError> {
-    DB.with(|f| f.execute("INSERT INTO dogs (url) VALUES (?1)", &[&image]))?;
-    Ok(())
-}
-
-// The database is only xavailable to server code
-#[cfg(feature = "server")]
-thread_local! {
-    pub static DB: rusqlite::Connection = {
-        // Open the database from the persisted "hotdog.db" file
-        let conn = rusqlite::Connection::open("hotdog.db").expect("Failed to open database");
-
-        // Create the "dogs" table if it doesn't already exist
-        conn.execute_batch(
-            "CREATE TABLE IF NOT EXISTS dogs (
-                id INTEGER PRIMARY KEY,
-                url TEXT NOT NULL
-            );",
-        ).unwrap();
-
-        // Return the connection
-        conn
-    };
-}
-use dioxus::prelude::*;
 
 #[derive(Routable, PartialEq, Clone)]
 enum Route {
@@ -118,44 +31,4 @@ enum Route {
     DogView,
     #[route("/favorites")]
     Favorites, // <------ add this new variant
-}
-
-
-
-#[component]
-pub fn Favorites() -> Element {
-    // Create a pending resource that resolves to the list of dogs from the backend
-    // Wait for the favorites list to resolve with `.suspend()`
-    let mut favorites = use_resource(list_dogs).suspend()?;
-    // let mut favorites = use_resource(super::backend::list_dogs).suspend()?;
-
-    rsx! {
-        div { id: "favorites",
-            div { id: "favorites-container",
-                for (id, url) in favorites().unwrap() {
-                    // Render a div for each photo using the dog's ID as the list key
-                    div {
-                        key: id,
-                        class: "favorite-dog",
-                        img { src: "{url}" }
-                    }
-                }
-            }
-        }
-    }
-}
-
-// Query the database and return the last 10 dogs and their url
-#[server]
-pub async fn list_dogs() -> Result<Vec<(usize, String)>, ServerFnError> {
-    let dogs = DB.with(|f| {
-        f.prepare("SELECT id, url FROM dogs ORDER BY id DESC LIMIT 10")
-            .unwrap()
-            .query_map([], |row| Ok((row.get(0)?, row.get(1)?)))
-            .unwrap()
-            .map(|r| r.unwrap())
-            .collect()
-    });
-
-    Ok(dogs)
 }
